+++ src/exactvalue.js
@@ -1,11 +1,11 @@
-import ValidationResult from './ValidationResult';
+import validator from './validator';
 import { isEqual } from 'lodash';
 
 export default function exactvalue(exact = 0, props = {}) {
     props.message = props.message || `Exactly ${exact}`;
 
-    return (value) => {
-        const isValid = !value || isEqual(value, exact);
-        return new ValidationResult(isValid, props);
-    };
+    return validator(
+        (value) => isEqual(value, exact),
+        props
+    );
 }