+++ src/maxlength.js
@@ -1,10 +1,11 @@
-import ValidationResult from './ValidationResult';
+import validator from './validator';
+import { lte } from 'lodash';
 
 export default function maxlength(max = 0, props = {}) {
     props.message = props.message || `At most ${max} characters`;
 
-    return (value) => {
-        const isValid = !value || value.length <= max;
-        return new ValidationResult(isValid, props);
-    };
+    return validator(
+        (value) => lte(value.length, max),
+        props
+    );
 }