+++ src/minvalue.js
@@ -1,10 +1,11 @@
-import ValidationResult from './ValidationResult';
+import validator from './validator';
+import { gte } from 'lodash';
 
 export default function minvalue(min = 0, props = {}) {
     props.message = props.message || `At least ${min}`;
 
-    return (value) => {
-        const isValid = !value || value >= min;
-        return new ValidationResult(isValid, props);
-    };
+    return validator(
+        (value) => gte(value, min),
+        props
+    );
 }