+++ test/validator.spec.js
@@ -1,7 +1,7 @@
 import expect from 'expect';
-import custom from '../src/custom';
+import { validator } from '../src';
 
-describe('custom', () => {
+describe('validator', () => {
     describe('for empty values', () => {
         [
             { value: null, testName: 'null' },
@@ -9,7 +9,7 @@ describe('custom', () => {
             { value: '', testName: 'empty string' }
         ].forEach(({ value, testName }) => {
             it(`it recognizes the value (${testName}) as valid`, () => {
-                const validate = custom(() => false);
+                const validate = validator(() => false);
 
                 const result = validate(value);
                 expect(result.isValid).toBe(true);
@@ -18,11 +18,11 @@ describe('custom', () => {
             it(`it does not call the validate function for value (${testName})`, () => {
                 let called = false;
 
-                const validator = () => {
+                const validationFunction = () => {
                     called = true;
                 };
 
-                const validate = custom(validator);
+                const validate = validator(validationFunction);
                 validate(value);
 
                 expect(called).toBe(false);
@@ -33,28 +33,28 @@ describe('custom', () => {
     describe('calls the validate function', () => {
         it('and passes the value', () => {
             let validatedValue;
-            const validator = (value) => {
+            const validationFunction = (value) => {
                 validatedValue = value;
                 return false;
             };
 
-            const validate = custom(validator);
+            const validate = validator(validationFunction);
             validate(4);
 
             expect(validatedValue).toBe(4);
         });
 
         it('and can set isValid to false', () => {
-            const validator = () => false;
-            const validate = custom(validator);
+            const validationFunction = () => false;
+            const validate = validator(validationFunction);
             const result = validate(true);
 
             expect(result.isValid).toBe(false);
         });
 
         it('and can set isValid to true', () => {
-            const validator = () => true;
-            const validate = custom(validator);
+            const validationFunction = () => true;
+            const validate = validator(validationFunction);
             const result = validate(false);
 
             expect(result.isValid).toBe(true);
@@ -62,24 +62,24 @@ describe('custom', () => {
 
         it('and passes the props', () => {
             let validatedProps;
-            const validator = (value, props) => {
+            const validationFunction = (value, props) => {
                 validatedProps = props;
             };
 
-            const props = { errorLevel: 10, message: 'Custom' };
-            const validate = custom(validator, props);
+            const props = { errorLevel: 10, message: 'validationFunction' };
+            const validate = validator(validationFunction, props);
             validate(true);
 
             expect(validatedProps).toEqual(props);
         });
 
         it('but does not let the function mutate the props', () => {
-            const validator = (value, props) => {
+            const validationFunction = (value, props) => {
                 props.mutated = true;
             };
 
             const props = { };
-            const validate = custom(validator, props);
+            const validate = validator(validationFunction, props);
             validate(true);
 
             expect(props.mutated).toNotExist();
@@ -88,13 +88,13 @@ describe('custom', () => {
         it('and uses a fresh copy of props for each call', () => {
             let mutated;
 
-            const validator = (value, props) => {
+            const validationFunction = (value, props) => {
                 mutated = props.mutated;
                 props.mutated = true;
             };
 
             const props = { };
-            const validate = custom(validator, props);
+            const validate = validator(validationFunction, props);
             validate(true);
 
             expect(mutated).toNotExist();
@@ -102,16 +102,16 @@ describe('custom', () => {
     });
 
     describe('message', () => {
-        const validator = () => false;
+        const validationFunction = () => false;
 
         it('defaults to "Invalid"', () => {
-            const validate = custom(validator);
+            const validate = validator(validationFunction);
             const result = validate('a');
             expect(result.message).toBe('Invalid');
         });
 
         it('can be overridden through props', () => {
-            const validate = custom(validator, { message: 'Overridden' });
+            const validate = validator(validationFunction, { message: 'Overridden' });
             const result = validate('a');
             expect(result.message).toBe('Overridden');
         });
@@ -119,7 +119,7 @@ describe('custom', () => {
 
     describe('props', () => {
         it('flow through', () => {
-            const validate = custom(() => false, { errorLevel: 10 });
+            const validate = validator(() => false, { errorLevel: 10 });
             const result = validate('a');
             expect(result.errorLevel).toBe(10);
         });